Major refactor of Controller.java
Controller.java had many of its functions moved to guiHelperMethods.java. As such, it is a lot easier to read/understand the class overall. 

It is still responsible for creating listeners, and calling the correct functions depending on what was clicked in the view.
